Move method for updating search results from App to Search component
