+++ src/components/Search.tsx
@@ -1,24 +1,24 @@
+// Import third party resources
 import { useState, useEffect } from 'react';
+import { useSelector, useDispatch } from "react-redux";
 import { Form, Button } from 'react-bootstrap';
 import { Typeahead } from 'react-bootstrap-typeahead';
 import 'react-bootstrap-typeahead/css/Typeahead.css';
 
-interface Props {
-    updateResults: any; 
-}
+// Import app resources
+import { RootState } from "../app/store";
+import { addLocation } from "../features/resultsSlice";
+import { LocationData, LocationCard } from "../types/types";
+import { formatNumber } from './Helpers';
 
-interface Location {
-	id: any;
-	location: any;
-	averageHomeValue: any;
-	averageRent: any;
-}
+const Search = () => {
 
-const Search = ( { updateResults }: Props) => {
+	const results = useSelector((state: RootState) => state.results.value)
+    const dispatch = useDispatch();
+
   	const [singleSelection, setSingleSelection] = useState([]);
 	const [data, setData] = useState([]);
 	
-
 	useEffect(() => {
 		async function fetchData() {
 			const response = await fetch('https://prhbo9omna.execute-api.us-east-1.amazonaws.com/getData');
@@ -32,6 +32,24 @@ const Search = ( { updateResults }: Props) => {
 		e.preventDefault();
 		updateResults(singleSelection);
 		setSingleSelection([])
+	}
+	
+	const updateResults = (search: LocationData[]) => {
+		const location = search[0].location;
+		const averageHomeValue = search[0].averageHomeValue;
+		const averageRent = search[0].averageRent ? search[0].averageRent : null;
+		const grossRentMultiplier = averageRent ? averageRent/averageHomeValue : null;
+		
+		const newLocationCard: LocationCard = {
+			id: results.length + 1,
+			location: location,
+			averageHomeValue: formatNumber('currency', averageHomeValue),
+			averageRent: averageRent ? formatNumber('currency', averageRent) : "Not Available",
+			grossRentMultiplier: grossRentMultiplier ? formatNumber('percent', grossRentMultiplier) : "Not Available"
+		
+		}
+
+		dispatch(addLocation(newLocationCard));
 	}
 
 	return (
